api/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
aiohttp>=3.9.0
aiofiles>=23.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
bcrypt==4.0.1
